Move communication service up
