+++ node/router/Cargo.toml
@@ -61,10 +61,6 @@ version = "1"
 path = "../../account"
 version = "=2.1.7"
 
-[dependencies.snarkos-node-sync-communication-service]
-path = "../sync/communication-service"
-version = "=2.1.7"
-
 [dependencies.snarkos-node-sync-locators]
 path = "../sync/locators"
 version = "=2.1.7"
