+++ node/src/client/router.rs
@@ -21,6 +21,7 @@ use snarkos_node_router::{
 use snarkos_node_tcp::{Connection, ConnectionSide, Tcp};
 use snarkvm::prelude::{block::Transaction, Network};
 
+use snarkos_node_sync::communication_service::CommunicationService;
 use std::{io, net::SocketAddr, time::Duration};
 
 impl<N: Network, C: ConsensusStorage<N>> P2P for Client<N, C> {
@@ -93,7 +94,7 @@ impl<N: Network, C: ConsensusStorage<N>> Reading for Client<N, C> {
     /// Creates a [`Decoder`] used to interpret messages from the network.
     /// The `side` param indicates the connection side **from the node's perspective**.
     fn codec(&self, _peer_addr: SocketAddr, _side: ConnectionSide) -> Self::Codec {
-        self.router.codec(_peer_addr, _side)
+        Default::default()
     }
 
     /// Processes a message received from the network.
@@ -102,12 +103,37 @@ impl<N: Network, C: ConsensusStorage<N>> Reading for Client<N, C> {
         if let Err(error) = self.inbound(peer_addr, message).await {
             if let Some(peer_ip) = self.router().resolve_to_listener(&peer_addr) {
                 warn!("Disconnecting from '{peer_ip}' - {error}");
-                self.send(peer_ip, Message::Disconnect(DisconnectReason::ProtocolViolation.into()));
+                Outbound::send(self, peer_ip, Message::Disconnect(DisconnectReason::ProtocolViolation.into()));
                 // Disconnect from this peer.
                 self.router().disconnect(peer_ip);
             }
         }
         Ok(())
+    }
+}
+
+#[async_trait]
+impl<N: Network, C: ConsensusStorage<N>> CommunicationService for Client<N, C> {
+    /// The message type.
+    type Message = Message<N>;
+
+    /// Prepares a block request to be sent.
+    fn prepare_block_request(start_height: u32, end_height: u32) -> Self::Message {
+        debug_assert!(start_height < end_height, "Invalid block request format");
+        Message::BlockRequest(BlockRequest { start_height, end_height })
+    }
+
+    /// Sends the given message to specified peer.
+    ///
+    /// This function returns as soon as the message is queued to be sent,
+    /// without waiting for the actual delivery; instead, the caller is provided with a [`oneshot::Receiver`]
+    /// which can be used to determine when and whether the message has been delivered.
+    async fn send(
+        &self,
+        peer_ip: SocketAddr,
+        message: Self::Message,
+    ) -> Option<tokio::sync::oneshot::Receiver<io::Result<()>>> {
+        Outbound::send(self, peer_ip, message)
     }
 }
 
@@ -155,7 +181,7 @@ impl<N: Network, C: ConsensusStorage<N>> Inbound<N> for Client<N, C> {
         }
 
         // Send a `Pong` message to the peer.
-        self.send(peer_ip, Message::Pong(Pong { is_fork: Some(false) }));
+        Outbound::send(self, peer_ip, Message::Pong(Pong { is_fork: Some(false) }));
         true
     }
